Move object module (non-breaking)
This moves the object module out of types and into
the location indicated in commit 20b26a5.

src/types/object.rs -> src/object.rs
